requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
numpy==1.24.3
Pillow==10.0.0
TensorFlow==2.13.0
scikit-learn==1.3.0
requests==2.31.0
matplotlib==3.7.2
